Add ApiEntityService, JSONApiParser and EntityCreator
The api entity service encapsulates communicating with the api,
parsing the response and creating the corresponding entities in the
local stores. For this it collaborates with an Api service, a parser
and an entity creator.

The parser is in charge of parsing the response and returning an object with two keys:
- __data__: The main data of the response. It can be a single item or a list of items.
- __included__ (optional): Additional parsed resources included with the response.

Each of this items must have two attributes:
- __type__: The type of the item that will be used by the creator
to obtain the corresponding store where to create the entity.
- __attributes__: The parsed attributes that will be passed to
the entity constructor when creating it.

In this commit the JSONApiParser is included that parses a JSONApi
response and returns the parsed response in the expected format.

The creator is in charge of creating the entities in their
corresponding stores. It receives an object with `data` and
`included` and for each item obtains the store for it from the `rootStore`
and calls the `create` method with the passed attributes in the obtained store.
Finally it returns the corresponding entity / entities for the data passed.